SAP bot/backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
langchain-text-splitters
PyPDF2
faiss-cpu
python-dotenv
pydantic
tiktoken
